scrape-to-s3/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
boto3==1.26.103
    # via
    #   -r requirements.in
    #   smart-open
botocore==1.29.103
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
idna==3.4
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mysql-connector-python==8.0.32
    # via -r requirements.in
protobuf==3.20.3
    # via mysql-connector-python
python-dateutil==2.8.2
    # via botocore
requests==2.28.2
    # via smart-open
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
smart-open[http,s3]==6.3.0
    # via -r requirements.in
urllib3==1.26.15
    # via
    #   botocore
    #   requests
